requirements.txt
flask
flask-cors
requests
web3
py-solc-x
eth-account
cdp-sdk
